endpointmanager: remove RemoveAll from interface
[ upstream commit e2fbd38eebcf1dfb27baeb36bded4a1ba2f4efc3 ]

Instead of having to import the testing pkg in non-testing code, let's
remove the blocker on just having this method in the _test.go files.

The endpoint manager in the daemon_test is re-initialized for every
test, and since we don't have package global state any more we can just
remove the call to RemoveAll, which solves all the problem nicely.
