+++ pkg/endpointmanager/allocator.go
@@ -6,8 +6,6 @@ package endpointmanager
 import (
 	"fmt"
 
-	"gopkg.in/check.v1"
-
 	"github.com/cilium/cilium/pkg/idpool"
 )
 
@@ -23,14 +21,6 @@ type epIDAllocator struct {
 func newEPIDAllocator() *epIDAllocator {
 	return &epIDAllocator{
 		pool: idpool.NewIDPool(minID, maxID),
-	}
-}
-
-// reallocatePool starts over with a new pool. This function is only used for
-// tests and its implementation is not optimized for production.
-func (a *epIDAllocator) reallocatePool(c *check.C) {
-	for i := uint16(minID); i <= uint16(maxID); i++ {
-		a.release(i)
 	}
 }
